requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
python-dotenv
pytest
pytest-mongodb
pytest-bdd
httpx
requests
motor
pandas
xgboost
scikit-learn
langchain_google_genai